pricing-service/requirements.txt
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
numpy>=1.24,<3.0
grpcio>=1.60,<2.0
grpcio-tools>=1.60,<2.0
grpcio-health-checking>=1.60,<2.0
